feat: added not found and bad request answers with req fields

+++ src/server.js
@@ -14,29 +14,59 @@ server.get('/videos', (req) => {
   return db.list(search)
 })
 
+server.get('/videos/:id', (req, res) => {
+  const video = db.get(req.params.id)
+
+  if (video.length)
+    return video
+
+  return res.status(404).send()
+})
+
 server.post('/videos', (req, res) => {
   const {title, description, duration} = req.body
-  db.create({
+  if (!title || !description || !duration) {
+    return res.status(400).send({ error: 'Fields title, description, and duration are required.' })
+  }
+  if (isNaN(duration))
+    return res.status(400).send({ error: 'Duration must be a number' })
+  const newVideo = db.create({
     title,
     description,
     duration
   })
-  return res.status(201).send()
+
+  return res.status(201).send(newVideo)
 })
 
 server.put('/videos/:id', (req, res) => {
   const {title, description, duration} = req.body
-  db.update(req.params.id, {
-    title,
-    description,
-    duration
-  })
-  return res.status(204).send()
+  if (!title || !description || !duration) {
+    return res.status(400).send({ error: 'Fields title, description, and duration are required.'})
+  }
+  const video = db.get(req.params.id)
+  if (video.length) {
+    db.update(req.params.id, {
+      title,
+      description,
+      duration
+    })
+  } else {
+    return res.status(404).send({ error: 'Not found any video within this id'})
+  }
+
+  return db.get(req.params.id)
 })
 
 server.delete('/videos/:id', (req, res) => {
-  db.delete(req.params.id)
-  return res.status(204).send()
+  const video = db.get(req.params.id)
+
+  if (video.length) {
+    db.delete(req.params.id)
+    return res.status(204).send()
+  } else {
+    return res.status(404).send()
+  }
 })
 
 server.listen({
